some mobigo reorganization
